day 4 part 1

+++ exercises/23-20-2.go
@@ -5,7 +5,7 @@ import (
 )
 
 func Year23Day20Exercise2() {
-	total := 0
+	total := int64(0)
 
 	fmt.Println(total)
 }
